bot-teams/requirements.txt
gunicorn
botbuilder-core
aiohttp
requests